Build file rewritten in Javascript to be run with Node. 'ninjagen.rb' replaced with 'setup.js'.
RubyInstaller switched packaged mingw-gcc toolchains the MSYS2, which made my atnivirus software prevent it from being downloaded as MSYS2 is detected as a virus by it.

Furthermore, the script needed an overhaul to add features needed by the project. Documentation for it will be added as the project progresses.

Also, there is still no project generation software out there that I like.