Implemented gc::collector::mark().
Expanded the documentation for gc::enumerate_cb.

No longer shrinks vectors during gc::collector::shrink() as that might try to allocate memory, which we must not do from the collector's worker.

Shrink the allocators vector to fit during init_impl() as ideally all allocators are tracked by that point and will not change.